game progress
each individual square was given an id and tag so that it can be easily identified. each square was also given alternating colors green and white (for now) with an if else statement to make sure that at the end of one row, the color that ends also starts the next so that it resembles a chess board.