added changing of players and flipping square IDs on the change of player
added a variable for the turn changing mechanic, and setting its text content to be brown at first. this means that square 0 should be top left, and after browns turn, the display should flip saying that its blacks turn and that square 0 is now bottom right. The call to change player happens after a dragOver event, and reverse and revertID functions are implemented using a fresh constant variable to grab all the squares and have them undergo a manipulation in their width in a loop to either reverse or revert the square ids.